Extracted branch advanced authoring to its own component and path. #2705
